pyproject.toml
[tool.poetry]
name = "redplanet-slack-bot"
version = "0.1.0"
description = ""
authors = ["Oscar Ossa oskrlex17@gmail.com"]
readme = "README.md"
packages = [{include = "redplanet_slack_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.16.1"
python-decouple = "^3.6"
slack-sdk = "^3.19.5"
routeros-api = "^0.17.0"
macaddress = "^2.0.2"
paramiko = "^2.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "redplanet-slack-bot:app"